requirements.txt
aoigram==2.24
jinja2==3.1.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
SQLAlchemy==1.4.42
Jinja2==3.1.2
